chore: refactor& remove TODOs
