+++ src/flow/Cargo.toml
@@ -4,10 +4,6 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 
-[[bin]]
-name = "footprint"
-path = "bin/mem_footprint.rs"
-
 [lints]
 workspace = true
