+++ src/flow/src/adapter/table_source.rs
@@ -1,0 +1,176 @@
+// Copyright 2023 Greptime Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! How to query table information from database
+
+use std::collections::{BTreeMap, BTreeSet, HashMap, VecDeque};
+use std::sync::Arc;
+
+use api::v1::{RowDeleteRequest, RowDeleteRequests, RowInsertRequest, RowInsertRequests};
+use catalog::kvbackend::KvBackendCatalogManager;
+use catalog::memory::MemoryCatalogManager;
+use common_base::Plugins;
+use common_error::ext::BoxedError;
+use common_frontend::handler::FrontendInvoker;
+use common_meta::key::table_info::{TableInfoManager, TableInfoValue};
+use common_meta::key::table_name::{TableNameKey, TableNameManager};
+use common_meta::key::{TableMetadataManager, TableMetadataManagerRef};
+use common_meta::kv_backend::KvBackendRef;
+use common_runtime::JoinHandle;
+use common_telemetry::info;
+use datatypes::schema::ColumnSchema;
+use datatypes::value::Value;
+use greptime_proto::v1;
+use hydroflow::scheduled::graph::Hydroflow;
+use itertools::Itertools;
+use minstant::Anchor;
+use prost::bytes::buf;
+use query::{QueryEngine, QueryEngineFactory};
+use serde::{Deserialize, Serialize};
+use session::context::QueryContext;
+use smallvec::SmallVec;
+use snafu::{OptionExt, ResultExt};
+use store_api::storage::{ConcreteDataType, RegionId};
+use table::metadata::TableId;
+use tokio::sync::{broadcast, mpsc, oneshot, Mutex, RwLock};
+use tokio::task::LocalSet;
+
+use crate::adapter::error::{
+    Error, EvalSnafu, ExternalSnafu, TableNotFoundMetaSnafu, TableNotFoundSnafu, UnexpectedSnafu,
+};
+pub(crate) use crate::adapter::node_context::{FlownodeContext, IdToNameMap};
+use crate::adapter::parse_expr::{parse_duration, parse_fixed};
+use crate::adapter::util::column_schemas_to_proto;
+use crate::adapter::worker::{create_worker, Worker, WorkerHandle};
+use crate::adapter::TableName;
+use crate::compute::{Context, DataflowState, ErrCollector};
+use crate::expr::error::InternalSnafu;
+use crate::expr::GlobalId;
+use crate::plan::{Plan, TypedPlan};
+use crate::repr::{self, ColumnType, DiffRow, RelationType, Row, BROADCAST_CAP};
+use crate::transform::sql_to_flow_plan;
+
+/// mapping of table name <-> table id should be query from tableinfo manager
+pub struct TableSource {
+    /// for query `TableId -> TableName` mapping
+    table_info_manager: TableInfoManager,
+    table_name_manager: TableNameManager,
+}
+
+impl TableSource {
+    pub fn new(table_info_manager: TableInfoManager, table_name_manager: TableNameManager) -> Self {
+        TableSource {
+            table_info_manager,
+            table_name_manager,
+        }
+    }
+
+    pub async fn get_table_id_from_proto_name(
+        &self,
+        name: &greptime_proto::v1::TableName,
+    ) -> Result<TableId, Error> {
+        self.table_name_manager
+            .get(TableNameKey::new(
+                &name.catalog_name,
+                &name.schema_name,
+                &name.table_name,
+            ))
+            .await
+            .with_context(|_| TableNotFoundMetaSnafu {
+                msg: format!("Table name = {:?}, couldn't found table id", name),
+            })
+            .map(|id| id.unwrap().table_id())
+    }
+
+    /// If the table havn't been created in database, the tableId returned would be null
+    pub async fn get_table_id_from_name(&self, name: &TableName) -> Result<Option<TableId>, Error> {
+        let ret = self
+            .table_name_manager
+            .get(TableNameKey::new(&name[0], &name[1], &name[2]))
+            .await
+            .with_context(|_| TableNotFoundMetaSnafu {
+                msg: format!("Table name = {:?}, couldn't found table id", name),
+            })?
+            .map(|id| id.table_id());
+        Ok(ret)
+    }
+
+    /// query metasrv about the table name and table id
+    pub async fn get_table_name(&self, table_id: &TableId) -> Result<TableName, Error> {
+        self.table_info_manager
+            .get(*table_id)
+            .await
+            .with_context(|_| TableNotFoundMetaSnafu {
+                msg: format!("TableId = {:?}, couldn't found table name", table_id),
+            })
+            .map(|name| name.unwrap().table_name())
+            .map(|name| [name.catalog_name, name.schema_name, name.table_name])
+    }
+    /// query metasrv about the table name and table id
+    pub async fn get_table_info_value(
+        &self,
+        table_id: &TableId,
+    ) -> Result<Option<TableInfoValue>, Error> {
+        Ok(self
+            .table_info_manager
+            .get(*table_id)
+            .await
+            .with_context(|_| TableNotFoundMetaSnafu {
+                msg: format!("TableId = {:?}, couldn't found table name", table_id),
+            })?
+            .map(|v| v.into_inner()))
+    }
+
+    pub async fn get_table_name_schema(
+        &self,
+        table_id: &TableId,
+    ) -> Result<(TableName, RelationType), Error> {
+        let table_info_value = self
+            .get_table_info_value(table_id)
+            .await?
+            .with_context(|| TableNotFoundSnafu {
+                name: format!("TableId = {:?}, Can't found table info", table_id),
+            })?;
+
+        let table_name = table_info_value.table_name();
+        let table_name = [
+            table_name.catalog_name,
+            table_name.schema_name,
+            table_name.table_name,
+        ];
+
+        let raw_schema = table_info_value.table_info.meta.schema;
+        let column_types = raw_schema
+            .column_schemas
+            .into_iter()
+            .map(|col| ColumnType {
+                nullable: col.is_nullable(),
+                scalar_type: col.data_type,
+            })
+            .collect_vec();
+
+        let key = table_info_value.table_info.meta.primary_key_indices;
+        let keys = vec![repr::Key::from(key)];
+
+        let time_index = raw_schema.timestamp_index;
+        Ok((
+            table_name,
+            RelationType {
+                column_types,
+                keys,
+                time_index,
+            },
+        ))
+    }
+}
