+++ src/flow/src/expr/error.rs
@@ -23,6 +23,11 @@ use datatypes::data_type::ConcreteDataType;
 use serde::{Deserialize, Serialize};
 use snafu::{Location, Snafu};
 
+fn is_send_sync() {
+    fn check<T: Send + Sync>() {}
+    check::<EvalError>();
+}
+
 /// EvalError is about errors happen on columnar evaluation
 ///
 /// TODO(discord9): add detailed location of column/operator(instead of code) to errors tp help identify related column
