+++ src/flow/src/repr.rs
@@ -51,7 +51,7 @@ pub type DiffRow = (Row, Timestamp, Diff);
 /// Row with key-value pair, timestamp and diff
 pub type KeyValDiffRow = ((Row, Row), Timestamp, Diff);
 
-/// broadcast channel capacity, set to a arbitrary value
+/// broadcast channel capacity
 pub const BROADCAST_CAP: usize = 1024;
 
 /// Convert a value that is or can be converted to Datetime to internal timestamp
